pyproject.toml
[tool.poetry]
name = "colibri"
version = "0.1.0"
description = ""
authors = ["Andres Jaimovich <andijaimo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.1"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
django-filter = "^24.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 100
exclude = '''
/(
  | migrations
)/

'''

[tool.isort]
profile = "black"
line_length = 100